requirements.txt
Django==1.8.6
MySQL-python==1.2.5
django-rest-swagger==0.3.5
djangorestframework==3.3.2
django-haystack==2.4.1
requests-aws4auth==0.9
elasticsearch==2.3.0
